// packages/text-engine/src/render/unicode-font-routing.win32.generated.ts
export interface Win32UnicodeFontEntry { file: string; postscriptName?: string }

export const UNICODE_FONT_FILES_WIN32: Record<string, Win32UnicodeFontEntry> = {
  "u-arial": { file: "arial.ttf" },
  "u-segoe-ui-historic": { file: "seguihis.ttf" },
  "u-mv-boli": { file: "mvboli.ttf" },
  "u-ebrima": { file: "ebrima.ttf" },
  "u-sans-serif-collection": { file: "SansSerifCollection.ttf" },
  "u-nirmala-ui": { file: "Nirmala.ttc", postscriptName: "NirmalaUI" },
  "u-tahoma": { file: "tahoma.ttf" },
  "u-leelawadee-ui": { file: "leelawui.ttf" },
  "u-microsoft-himalaya": { file: "himalaya.ttf" },
  "u-myanmar-text": { file: "mmrtext.ttf" },
  "u-sylfaen": { file: "sylfaen.ttf" },
  "u-malgun-gothic": { file: "malgun.ttf", postscriptName: "MalgunGothic" },
  "u-gadugi": { file: "gadugi.ttf" },
  "u-mongolian-baiti": { file: "monbaiti.ttf" },
  "u-microsoft-tai-le": { file: "taile.ttf" },
  "u-microsoft-new-tai-lue": { file: "ntailu.ttf" },
  "u-segoe-ui": { file: "segoeui.ttf" },
  "u-times-new-roman": { file: "times.ttf" },
  "u-segoe-ui-symbol": { file: "seguisym.ttf" },
  "u-cambria-math": { file: "cambria.ttc", postscriptName: "CambriaMath" },
  "u-lucida-sans-unicode": { file: "l_10646.ttf" },
  "u-ms-pgothic": { file: "msgothic.ttc", postscriptName: "MS-PGothic" },
  "u-microsoft-sans-serif": { file: "micross.ttf" },
  "u-microsoft-yahei": { file: "msyh.ttc", postscriptName: "MicrosoftYaHei" },
  "u-simsun": { file: "simsun.ttc", postscriptName: "SimSun" },
  "u-yu-gothic": { file: "YuGothR.ttc", postscriptName: "YuGothic-Regular" },
  "u-microsoft-jhenghei-ui": { file: "msjh.ttc", postscriptName: "MicrosoftJhengHeiUIRegular" },
  "u-yu-gothic-ui": { file: "YuGothR.ttc", postscriptName: "YuGothicUI-Semilight" },
  "u-microsoft-yi-baiti": { file: "msyi.ttf" },
  "u-microsoft-phagspa": { file: "phagspa.ttf" },
  "u-javanese-text": { file: "javatext.ttf" },
  "u-calibri": { file: "calibri.ttf" },
  "u-simsun-extb": { file: "simsunb.ttf" },
  "u-simsun-extg": { file: "SimsunExtG.ttf" },
};